feat: небольшие правки
