+++ index.html
@@ -12,12 +12,15 @@
 <body class="body">
   <!-- Page -->
   <div class="page">
+
     <!-- Header -->
     <header class="header section section_size_small page__container">
       <img class="header__logo" src="./images/mesto-logo.svg" alt="Логотип Место">
     </header>
     <!-- /Header -->
+
     <main class="main page__container">
+
       <!-- Section Profile -->
       <section class="profile">
         <div class="profile__user-info">
@@ -33,12 +36,14 @@
         <button class="profile__add-button" type="button" name="addButton"></button>
       </section>
       <!-- /Section Profile -->
+
       <!-- Section Elements -->
       <section class="elements">
         <ul class="elements__element-list"></ul>
         <span class="elements__msgNoElements">Нет элементов</span>
       </section>
       <!-- /Section Elements -->
+
     </main>
 
     <!-- Popup UserProfile-->
@@ -110,9 +115,9 @@
     </li>
   </template>
   <!-- /Template for block Element -->
-  <script defer src="./scripts/validate.js"></script>
+  <!-- <script defer src="./scripts/validate.js"></script> -->
   <script defer src="./scripts/index.js" type="module"></script>
-  <script defer src="./scripts/Card.js" type="module"></script>
+  <!-- <script defer src="./scripts/Card.js" type="module"></script> -->
 </body>
 
 </html>
